+++ packages/iov-lisk/src/liskconnector.ts
@@ -3,7 +3,7 @@ import { ChainConnector, ChainId } from "@iov/bcp";
 import { liskCodec } from "./liskcodec";
 import { LiskConnection } from "./liskconnection";
 
-export function createLiskConnector(url: string, expectedChainId?: ChainId): ChainConnector {
+export function createLiskConnector(url: string, expectedChainId?: ChainId): ChainConnector<LiskConnection> {
   return {
     establishConnection: async () => LiskConnection.establish(url),
     codec: liskCodec,