+++ packages/iov-rise/types/riseconnector.d.ts
@@ -1,2 +1,6 @@
 import { ChainConnector, ChainId } from "@iov/bcp";
-export declare function createRiseConnector(url: string, expectedChainId?: ChainId): ChainConnector;
+import { RiseConnection } from "./riseconnection";
+export declare function createRiseConnector(
+  url: string,
+  expectedChainId?: ChainId,
+): ChainConnector<RiseConnection>;